move event blueprints from dev to stanford
